use malloc to allocate per-thread lwip state on demand. still need a mechanism to GC it.
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@467 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ lib/lwip/jos64/arch/sys_arch.h
@@ -5,7 +5,7 @@
 
 typedef	int sys_sem_t;
 typedef int sys_mbox_t;
-typedef int sys_thread_t;
+typedef uint64_t sys_thread_t;
 
 #define SYS_MBOX_NULL	(-1)
 #define SYS_SEM_NULL	(-1)
